Fix retries overflow, jitter NaN/boundary bugs; add fuzz tests
Bounds-checking review (prompted by hand-picked edge cases, then
confirmed with real fuzzing) found three concrete bugs:

- State.Retries overflow: incrementing at the max representable value
  panicked in Rust debug builds and silently wrapped to a negative
  number in Go, which would disable the MaxRetries termination check
  (a negative Retries is never greater than any sane positive
  MaxRetries). Fixed with a new saturatingAddInt helper in Go, and by
  switching to the built-in u64::saturating_add in Rust.

- Go's applyJitter overflow guard had a float-rounding hole:
  float64(math.MaxInt64) rounds up to exactly 2^63 (a float64 mantissa
  can't hold all 63 significant bits), so the 'rtF > float64(MaxInt64)'
  check silently failed at the boundary, and converting the
  out-of-range float produced an implementation-defined result -- in
  practice, a huge *negative* Duration. Fixed by using >= instead of >.

- NaN jitter propagated into undefined (Go) or surprising-but-defined
  (Rust) results, including via a 0 * Infinity edge case when base is
  zero. Both languages now detect a NaN result and fall back to the
  unjittered base, documented in SPEC.md section 5.2 as 'NaN jitter
  behaves like 0.0 jitter'.

Also, per discussion: widened Rust's State.retries / Params.max_retries
from u32 to u64 to match Go's practically-unbounded int, rather than
just saturating a narrower type.

Added regression tests for all three bugs, plus new property/fuzz test
infrastructure to catch the next one automatically instead of by hand:

- go/fuzz_test.go: native Go fuzzing (testing.F, stdlib since 1.18, no
  new dependency), seeded with every edge case found above. Run for
  real: ~11M executions across two runs, zero failures.
- rust/tests/properties.rs: proptest (new dev-dependency only). Run
  with 100k+ cases, zero failures.

Both check: compute never panics, Elapsed never decreases, Retries
increments correctly or saturates, the termination decision matches
SPEC.md section 5.4's order, and NaN jitter equals zero jitter.

make check / CI only run the fast seeded/default-case versions; new
'make go-fuzz' and 'make rust-proptest-deep' targets run deeper local
searches on demand.

Documented Rust's remaining known limitation (apply_jitter's ceiling is
~584 years, short of Duration::MAX's ~584 billion) as an accepted f64
precision boundary rather than a bug, in SPEC.md section 5.1.

SPEC.md updated throughout: section 2 (repo tree), 4.1 (Rust retries
type), 5.2 (NaN rule), 5.3 (Retries saturation), 8.1 (Rust API note),
and a new section 13 describing the testing strategy.

All validated via 'make check' (go vet/fmt/staticcheck/test, cargo
fmt/clippy/test with and without the rand feature) plus the extended
fuzz/proptest runs noted above.

+++ rust/tests/conformance.rs
@@ -56,7 +56,7 @@ impl VectorParams {
         Params {
             initial_rt: ms(self.initial_rt_ms),
             max_interval: self.max_interval_ms.map(ms),
-            max_retries: self.max_retries.map(|n| n as u32),
+            max_retries: self.max_retries.map(|n| n as u64),
             max_duration: self.max_duration_ms.map(ms),
         }
     }
